melon_util/pyproject.toml
[tool.poetry]
name = "melon-util"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydub = "^0.25.1"
loguru = "^0.7.2"
pillow = "^10.0.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
ruff = "^0.0.290"
mypy = "^1.5.1"
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
show_error_context = true
show_column_numbers = true
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
exclude = ["^cdk.out/$"]

[tool.ruff]
target-version = "py311"
select = ["ALL"]
ignore = [
  "D",
]
fixable = ["ALL"]
unfixable = []
line-length = 88
